refactor: displayname 수정
