--- src/main/java/domain/WinnerJudge.java
@@ -1,7 +1,0 @@
-package domain;
-
-import java.util.List;
-
-public interface WinnerJudge {
-    public List<Car> getWinner(List<Car> cars);
-}
