+++ src/main/java/racing/Application.java
@@ -1,3 +1,5 @@
+package racing;
+
 public class Application {
     public static void main(String[] args) {
         RaceController controller = new RaceController();
