+++ src/main/java/racing/ExceptionHandlingTemplate.java
@@ -1,12 +1,12 @@
-package domain;
+package racing;
 
 import java.util.function.Function;
 import java.util.function.Supplier;
-import view.OutputView;
+import racing.view.OutputView;
 
 public interface ExceptionHandlingTemplate {
     OutputView outputView  = new OutputView();
-    static <T, R> R repeatUntilReadValidInput(Function<T, R> object, Supplier<T> input) {
+    static  <T, R> R repeatUntilReadValidInput(Function<T, R> object, Supplier<T> input) {
         while (true) {
             try {
                 return object.apply(input.get());
