+++ src/main/java/racing/domain/car/CarDto.java
@@ -1,4 +1,4 @@
-package domain;
+package racing.domain.car;
 
 public class CarDto {
     private final String name;
