+++ src/main/java/racing/domain/race/NumberPicker.java
@@ -1,4 +1,4 @@
-package domain;
+package racing.domain.race;
 
 public interface NumberPicker {
     int pickNumber();
