+++ src/main/java/racing/domain/race/WinnerJudge.java
@@ -1,0 +1,8 @@
+package racing.domain.race;
+
+import racing.domain.car.Car;
+import java.util.List;
+
+public interface WinnerJudge {
+    List<Car> getWinner(List<Car> cars);
+}
