+++ src/test/java/domain/CarTest.java
@@ -1,5 +1,6 @@
 package domain;
 
+import racing.domain.car.Car;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
